+++ js/result/result-view.js
@@ -1,12 +1,12 @@
 import AbstractView from '../view';
 
-const drawComparison = (data) => {
-  let comparison = ``;
-  if (data.comparison) {
-    comparison = `<span class="main-comparison">${data.comparison}</span>`;
-  }
-  return comparison;
-};
+// export const drawComparison = (data) => {
+//   let comparison = ``;
+//   if (data.comparison) {
+//     comparison = `<span class="main-comparison">${data.comparison}</span>`;
+//   }
+//   return comparison;
+// };
 
 export default class ResultView extends AbstractView {
   constructor(data) {
@@ -19,7 +19,7 @@ export default class ResultView extends AbstractView {
     <section class="logo" title="Угадай мелодию"><h1>Угадай мелодию</h1></section>
     <h2 class="title">${this.data.title}</h2>
     <div class="main-stat">${this.data.statistics}</div>
-    ${drawComparison(this.data)}
+    <span class="main-comparison">${this.data.comparison ? this.data.comparison : ``}</span>
     <span role="button" tabindex="0" class="main-replay">${this.data.replayTitle}</span>
   </section>`;
   }
