+++ js/winscreen/winscreen-view.js
@@ -1,0 +1,8 @@
+import ResultView from '../result/result-view';
+
+export default class WinScreenView extends ResultView {
+  constructor(data) {
+    super();
+    this.data = data;
+  }
+}
